pyproject.toml
[tool.poetry]
name = "all-crosscode-versions-expansion"
version = "0.0.0"
description = ""
authors = []
license = "CC0-1.0"

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.4.1"
steam = {extras = ["client"], version = "^1.2.0"}
typing-extensions = "^3.10.0"
mwclient = "^0.10.1"

[tool.poetry.dev-dependencies]
pyright = "^0.0.10"
yapf = "^0.31.0"
toml = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
venvPath = "."
venv = ".venv"
typeCheckingMode = "basic"

[tool.yapf]
based_on_style = "google"
column_limit = 99
indent_width = 2
continuation_indent_width = 2
dedent_closing_brackets = true
coalesce_brackets = true
spaces_around_power_operator = true
